pipeline/images/serving_image/requirements.txt
fastapi == 0.108.0
uvicorn[standard] == 0.27.0
google-cloud-storage == 2.11.0
pandas == 2.1.1
numpy == 1.26.0
scikit-learn == 1.3.1
xgboost == 2.0.0